Changes templating to use new configuration templating
